requirements.txt
flask==3.0.0
pillow==10.1.0
pandas==2.1.3
openpyxl==3.1.2
python-docx==1.1.0
twilio==8.10.0
python-dotenv==1.0.0
reportlab==4.0.7
PyPDF2==3.0.1
xlrd==2.0.1
